Update base.html
In the `pelicanconf.py` example in the `docs` directory, I found:
```
MENUITEMS = (
    ("Archives", "/archives.html"),
    ("Categories", "/categories.html"),
    ("Tags", "/tags.html"),
)
```
But this is not robust, as the site may be published to a URL with path components. So `SITEURL` should be prefixed to these URLs (and the trailing slash in the config could therefore be removed).
Granted, this makes it now impossible to put absolure URLs in `MENUITEMS`, but arguably, absolute URLs to external sites should rather be in LINKS.

The same logic applies to `SITELOGO`, so I updated it also.